refactor: 提取筛选组件
